requirements.txt
install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
ultralytics
dxcam
pyserial
pykdtree
